requirements.txt
beautifulsoup4==4.14.3
celery==5.6.3
fastapi==0.135.3
httpx==0.28.1
jieba==0.42.1
pillow==11.3.0
playwright==1.54.0
psycopg[binary]==3.3.3
redis==7.4.0
requests==2.33.1
uvicorn==0.44.0
websockets==16.0
